requirements.txt
pandas==2.3.1
psycopg==3.2.9
python-dotenv==1.1.1
timescale-vector==0.0.7
instructor==1.10.0
anthropic==0.59.0
sentence-transformers==5.0.0
datasets==4.0.0
